Fix resource fetch to plural form, otherwise this fails
Api resources are plural, when using the kind value to understand the
resources just add 's' for plural.

context: when fetching remediation objects we don't know have the
resource name but only the template kind (i.e
PoisonPillRemediationTemplate) so the resource must be
poisonpillremediations.
This can work but RBAC rules usually specify the resource
name, which is plural and not singular form.

Change-Id: Iae06668964d3c779a7f320c16e3e67389e68e3c5
Signed-off-by: Roy Golan <rgolan@redhat.com>