Make NHC operational on start
MOTIVATION
Make NHC operational on start without user actions.

MODIFICATION
Now that NHC is installed by default with poison-pill we make it
operational by creating a default worker-only NHC object that points to
a poison-pill remediation template.
Do note that the reconciler doesn't validate if a remediation template
by the supplied GVK exists. It only uses the GVK when actually creating
the remediation CR.
The responsibility of creating the default expected poison-pill template
is on the poison-pill operator.

RESULT
The opeartor will create an NHC CR for workers only with the defualt
spec parameters, and some default poison-pill remediation template.
If there is some NHC CR that exists then nothing changes - to support
upgrades/restarts.

Change-Id: I972a96985d41804f5e489fbe83055ff9c7d0d35a
Signed-off-by: Roy Golan <rgolan@redhat.com>