config/ fixes to make bundle generation right
- bundle now have service account (non 'default')
- cluster roles and binding of manager, aggregator, leader election

The main reason the files were not created by `make bundle` is because
operator-sdk silently failed because 1. the name of the subject in role
binding pointed to 'default' instead of 'controller-manager' 2. on the
new service_account.yaml specifying a namespace also cause the sdk to
ignore the file and silently skip generating the rest of the files.

Change-Id: I9d5239ea3db8ca2bbb3a7f51595641ba4ea19b58
Signed-off-by: Roy Golan <rgolan@redhat.com>