Support for pausing remediations per resource
Motivation
In some situation pausing the remediation is better than just removing
the CR altogether. It can express some period of time that the cluster
may go under changes or stress or just for investigations, and we want
to allow a temporary pause.

Modification
Introduce new spec field 'PauseRequests []string'
This API addition allows any party to add its request to a list
on the NHC resource and prevent any new remdiation while the request is
there.

Signed-off-by: Roy Golan <rgolan@redhat.com>