Apply pod security changes
In order to comply with the restricted profile, the pod needs to have
runAsNonRoot: true, and all containers need to have
allowPrivilegeEscalation:false and drop all caps.
Also the Dockerfile should set a USER.

On k8s we would also need to set seccompProfile.type: RuntimeDefault.
That would break deployment on OCP 4.10 though. So leave it empty,
which is fine also on OCP 4.11+

Signed-off-by: Marc Sluiter <msluiter@redhat.com>