Add Node providerID to metric
Add Node providerID to NodeHealthCheckUnhealthyNodeDuration metric.

While node name might change after a remediation that deletes the node
(e.g. with Machine-Deletion-Remediation), the ProviderID is expected to
be the same, allowing users to query for the complete history of the
same node after such remediations.

NOTE: ProviderID is expected to change anyway in case of Cloud cluster
providers.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>