Add bundle Dockerfile for CI E2E tests
Currently, the console-plugin excluded from the bundle, since we
committed the k8s one, and this means CI E2E tests cannot test it.

To keep the k8s bundle committed and also test all NHC components, this
change adds a bundle.ci.Dockerfile which creates the OCP bundle version,
which contains the console-plugin.

For this to work, CI needs to use this "special" Dockerfile in place of
the usual bundle.Dockerfile.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>