Prevent overriding of pullspec set by CI
CI is expected to set a specific pullspec for the NHC image it is
building and testing. This pullspec however was overridden by our
multistep image build in bundle.ci.Dockerfile (which in turn is
necessary to test the console-plugin, which is only contained in the OCP
bundle).

This change prevents such overriding by:
- getting the current pullspec, set by CI, and
- exporting it as IMG environ variable to be used when building the
  bundle.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>