Fix Go version used in Dockerfile to build the binary
The Dockerfile currently installs the Go version defined in go.mod using
go get. However, this method doesn't ensure that the just-installed Go
version is the same one used to build the binary. Instead, the binary is
still built with the Go version released with the container.

To address this issue, this change downloads and extracts the Go archive
corresponding to the go.mod file and sets the PATH to prioritize the
current installation.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>