Fix orphaned CR handling
No remediator is deleting AND recreating nodes anymore, so it's safe
to delete the CR when its node was deleted.
For backwards compapibilty we keep the old behaviour for remediators
which mark their CR with the PermanentNodeDeletionExpected condition:
the CR will only be deleted once it succeeded.

Fixes the case when escalating remediation with MDR was used, and
CRs for other remediators weren't cleaned up after node deletion.

Signed-off-by: Marc Sluiter <msluiter@redhat.com>